+++ include/PRadGEMFEC.h
@@ -45,7 +45,7 @@ public:
     PRadGEMSystem *GetSystem() const {return gem_srs;};
     int GetID() const {return id;};
     const std::string &GetIP() const {return ip;};
-    size_t GetCapacity() const {return adc_list.size();};
+    uint32_t GetCapacity() const {return adc_list.size();};
     PRadGEMAPV *GetAPV(const int &slot) const;
     std::vector<PRadGEMAPV*> GetAPVList() const;
